Added filter x64 support and removed all warnings

+++ directshow/Lightpack-filter/CLightpack.cpp
@@ -374,7 +374,7 @@ void CLightpack::postConnection()
     }
 
     // Update the size of the color array
-    int numLeds = mDevice->getCountLeds();
+    size_t numLeds = mDevice->getCountLeds();
     if (mPropColors.size() != numLeds) {
         mPropColors.resize(numLeds);
     }
